BACK-1009 redefine name events to event and add new attributes in model event
